Refactor: Arrange snipperts router
